analysis-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
aiofiles==24.1.0
openpyxl==3.1.5
httpx==0.27.0
python-docx==1.1.2
PyPDF2==3.0.1
matplotlib==3.9.1
reportlab==4.2.2
